requirementss.txt
Django==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
whitenoise==6.6.0
gunicorn==21.2.0
python-dateutil==2.9.0